* Refactor Templates / Template Folders for proper internal / public classes and files *
